# fix(getAssignedEntityChannels): fix performance problems in v2.2