@michaelbromley 2024-04-25
Thanks for the contribution. I suspect that this may be solving the same issue as

- https://github.com/vendure-ecommerce/vendure/pull/2791

Can you let me know what you think? @monrostar can you weigh in?