@alexisvigoureux 2024-04-25
Thx for your answers, @monrostar your PR is better, I will close mine